Move content of src/docs -> docs
