+++ cmd/singularity/cli/capability_list.go
@@ -9,7 +9,7 @@ package cli
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/sylabs/singularity/src/docs"
+	"github.com/sylabs/singularity/docs"
 )
 
 func init() {
