+++ cmd/singularity/cli/instance_stop.go
@@ -10,7 +10,7 @@ package cli
 import (
 	"errors"
 	"github.com/spf13/cobra"
-	"github.com/sylabs/singularity/src/docs"
+	"github.com/sylabs/singularity/docs"
 )
 
 func init() {
