+++ cmd/singularity/cli/keys.go
@@ -8,7 +8,7 @@ package cli
 import (
 	"errors"
 	"github.com/spf13/cobra"
-	"github.com/sylabs/singularity/src/docs"
+	"github.com/sylabs/singularity/docs"
 )
 
 const (
