+++ cmd/singularity/cli/keys_list.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/sylabs/singularity/pkg/sypgp"
-	"github.com/sylabs/singularity/src/docs"
+	"github.com/sylabs/singularity/docs"
 )
 
 var secret bool
