+++ cmd/singularity/cli/keys_push.go
@@ -13,7 +13,7 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/sylabs/singularity/internal/pkg/sylog"
 	"github.com/sylabs/singularity/pkg/sypgp"
-	"github.com/sylabs/singularity/src/docs"
+	"github.com/sylabs/singularity/docs"
 )
 
 func init() {
