+++ cmd/singularity/cli/pull.go
@@ -7,7 +7,7 @@ package cli
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/sylabs/singularity/src/docs"
+	"github.com/sylabs/singularity/docs"
 )
 
 const (
