+++ mlocal/frags/Makefile.stub
@@ -226,7 +226,7 @@ $(actions_INSTALL): $(actions_LIST)
 
 .PHONY: man
 man:
-	$(V)go run $(GO_TAGS) $(SOURCEDIR)/src/docs/makeDocs/makeDocs.go $(DESTDIR)$(MANDIR)/man1
+	$(V)go run $(GO_TAGS) $(SOURCEDIR)/docs/makeDocs/makeDocs.go $(DESTDIR)$(MANDIR)/man1
 
 .PHONY: collect
 collect:
